Fix build failures from library changes
ran into some dependency issues building, looks like in the last few years some
libraries have changed.

fixes for the following exiv2 api chanegs:
- toLong replaced with toInt64 https://github.com/Exiv2/exiv2/pull/2062
- AnyError renamted to Error https://github.com/Exiv2/exiv2/pull/2141

downgraded lensfun 0.3.95 -> 0.3.4. This is more of an opinionated change but
it's probably best to use lensfun 0.3.4, they never shipped a stable version of
0.3.95 and in the meantime the rest of the ecosystem has settled on 0.3.4 with
continual updates to the database. For what it's worth rawtherapee, digikam, and
darktable all link against 0.3.4 and every distro ships 0.3.4 with the latest
lens db.

haven't made the appropriate changes to CI but it looks like this is already
failing for other reasons